Change input type from int to float

+++ Atividades 1 a 15/4 - Dobro, Triplo e Metade.py
@@ -1,3 +1,3 @@
-n = int(input("Digite um número: "))
+n = float(input("Digite um número: "))
 
 print(f" Dobro: {n*2}\n Triplo: {n*3}\n Metade: {n/2}")
